requirement.txt
fastapi==0.100.0
openai==1.3.0
pydantic==2.0.2
python-dotenv==1.0.0
requests==2.31.0
transformers==4.35.2
soundfile==0.12.1
torch==2.1.1
huggingface-hub==0.19.4
python-multipart==0.0.6
pydub==0.25.1
numpy==1.24.3